moves original example server into server-nuxtjs
