+++ server-nuxtjs/server/middleware/index.js
