+++ server-nuxtjs/server/middleware/installClientServerProxy.js
