+++ server-nuxtjs/server/middleware/installLogging.js
