+++ server-nuxtjs/server/middleware/installPassportStrategy.js
