+++ server-nuxtjs/server/middleware/installPostGraphile.js
