+++ server-nuxtjs/server/middleware/installSession.js
