+++ server-nuxtjs/server/middleware/installSharedStatic.js
